Clean Code
merge file to clean code
