--- Command/RelayCommand.cs
@@ -1,35 +1,0 @@
-﻿using System;
-using System.Windows.Input;
-
-namespace GoninDigital.Command
-{
-    class RelayCommand<T> : ICommand
-    {
-        private readonly Predicate<T> _canExecute;
-        private readonly Action<T> _execute;
-
-        public RelayCommand(Predicate<T> canExecute, Action<T> execute)
-        {
-            if (execute == null)
-                throw new ArgumentNullException("execute");
-            _canExecute = canExecute;
-            _execute = execute;
-        }
-
-        public bool CanExecute(object parameter)
-        {
-            return _canExecute == null ? true : _canExecute((T)parameter);
-        }
-
-        public void Execute(object parameter)
-        {
-            _execute((T)parameter);
-        }
-
-        public event EventHandler CanExecuteChanged
-        {
-            add { CommandManager.RequerySuggested += value; }
-            remove { CommandManager.RequerySuggested -= value; }
-        }
-    }
-}
