+++ ViewModels/BaseViewModel.cs
@@ -3,8 +3,10 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Linq;
 using System.Runtime.CompilerServices;
+using System.Security.Cryptography;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows.Input;
 
 namespace GoninDigital.ViewModels
 {
@@ -15,6 +17,41 @@ namespace GoninDigital.ViewModels
         protected virtual void OnPropertyChanged([CallerMemberName] string propertyName = null)
         {
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+        }
+    }
+    class RelayCommand<T> : ICommand
+    {
+        private readonly Predicate<T> _canExecute;
+        private readonly Action<T> _execute;
+
+        public RelayCommand(Predicate<T> canExecute, Action<T> execute)
+        {
+            if (execute == null)
+                throw new ArgumentNullException("execute");
+            _canExecute = canExecute;
+            _execute = execute;
+        }
+
+        public bool CanExecute(object parameter)
+        {
+            try
+            {
+                return _canExecute == null ? true : _canExecute((T)parameter);
+            }
+            catch
+            {
+                return true;
+            }
+        }
+        public void Execute(object parameter)
+        {
+            _execute((T)parameter);
+        }
+
+        public event EventHandler CanExecuteChanged
+        {
+            add { CommandManager.RequerySuggested += value; }
+            remove { CommandManager.RequerySuggested -= value; }
         }
     }
 }
