+++ ViewModels/LoginViewModel.cs
@@ -7,23 +7,15 @@ using System.Windows;
 using System.Windows.Input;
 using System.ComponentModel;
 using System.Windows.Controls;
-
 using GoninDigital.Models;
-using GoninDigital.Command;
 using GoninDigital.Utils;
-using GoninDigital.Views;
 
 namespace GoninDigital.ViewModels
 {
-    class LoginViewModel : INotifyPropertyChanged
+    class LoginViewModel : BaseViewModel
     {
-        Window curWindow;
-        public event PropertyChangedEventHandler PropertyChanged;
-        protected virtual void OnPropertyChanged(string propertyName)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
         #region Properties
+        Window curWindow;
         public Action CloseAction { get; set; }
         private string _Usrname;
         public string UserName
@@ -31,7 +23,6 @@ namespace GoninDigital.ViewModels
             get { return _Usrname; }
             set
             {
-                if (_Usrname == value) { return; }
                 _Usrname = value;
                 OnPropertyChanged(UserName);
             }
@@ -42,7 +33,6 @@ namespace GoninDigital.ViewModels
             get { return _Passwrd; }
             set
             {
-                if (_Passwrd == value) { return; }
                 _Passwrd = value;
                 OnPropertyChanged(Password);
             }
@@ -50,6 +40,7 @@ namespace GoninDigital.ViewModels
         public ICommand LoginCommand { get; set; }
         public ICommand RegisterCommand { get; set; }
         public ICommand LoadedLoginCommand { get; set; }
+        public ICommand PasswordChangedCommand { get; set; }
         #endregion
 
         #region Constructor
@@ -57,36 +48,34 @@ namespace GoninDigital.ViewModels
         {
             LoadedLoginCommand = new RelayCommand<Window>((p)=> { return true; }, (p)=> {
                 curWindow = p;
-                //p.Hide();
-                //Window ps = new();
-                //ps.ShowDialog();
-                //p.Show();
             });
             LoginCommand = new RelayCommand<Window>((p) => { return true; }, (p) => { LoginCommandExecute(curWindow); });
-            //RegisterCommand = new RelayCommand<Window>((p) => { return true; }, (p) => { RegisterCommandExcute(); });
+            PasswordChangedCommand = new RelayCommand<PasswordBox>((p) => { return true; }, (p) => { Password = p.Password; });
+            RegisterCommand = new RelayCommand<Window>((p) => { return true; }, (p) => { RegisterCommandExcute(); });
         }
         #endregion
 
         #region Private Methods
+        
         private void LoginCommandExecute(Window p)
         {
             _Usrname = UserName;
             if (_Usrname == null || _Passwrd == null)
             {
-                //MessageBox.Show("Both username and password should be filled in.");
+                MessageBox.Show("Both username and password should be filled in.");
             }
+
+            string passEncode = Encode.MD5Hash(Encode.Base64Encode(_Passwrd));
             int accCount = DataProvider.Instance.Db.Users.Where(x => x.UserName == _Usrname && x.Password == _Passwrd).Count();
             if (accCount > 0)
             {
                 var homepageViewModel = new MainWindow();
                 homepageViewModel.Show();
-                curWindow.Close();
-
-                //WindowManager.ChangeWindowContent(p, homepageViewModel, "Homepage GoninDigital", "GoninDigital.Views.DashboardView");
+                p.Close();
             }
             else
             {
-                //MessageBox.Show("Invalid credentials.");
+                MessageBox.Show("Invalid credentials.");
             }
         }
         private void RegisterCommandExcute()
