+++ ViewModels/MainViewModel.cs
@@ -5,7 +5,6 @@ using System.Text;
 using System.Threading.Tasks;
 using System.ComponentModel;
 using System.Windows;
-using GoninDigital.Command;
 using GoninDigital.Models;
 using GoninDigital.Views;
 using System.Windows.Input;
